% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie.r
\name{hazard}
\alias{hazard}
\title{Calculates the hazard for a stochastic chemical kinetic system}
\usage{
hazard(sys, tau = 1, engine = "R")
}
\arguments{
\item{sys}{the stochastic chemical kinetic system}

\item{tau}{the time interval for the hazard}

\item{engine}{use 'R' or 'C' code}
}
\value{
a list containing the hazard and the hazard part (hazard divided by rates)
}
\description{
Calculates the hazard for a stochastic chemical kinetic system
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\seealso{
\code{\link{tau_leap}}
}

